Refactor : 클라이언트 Socket class 파일명 index.ts로 수정
